requirements.txt
numpy
pandas
matplotlib
streamlit
streamlit-vis-timeline>0.2.0
pyxlsb
psycopg2-binary==0.9.9
xlsxwriter